expansion & idk what changed
the commit "idk what this is lemme check" was some unlabeled changes i found on git bash but not git desktop. it didnt look like they changed anything but i committed it just in case cause i know nik was doing something on my computer. if anything is broken lemme know and ill fix the bash